src/handler/phone-number/remove-phone-number.test.ts
import { ClientError } from "@lindorm-io/errors";
import { PhoneNumber } from "../../entity";
import { removePhoneNumber } from "./remove-phone-number";
import { logger } from "../../test";

describe("removePhoneNumber", () => {
  const phone1 = new PhoneNumber({
    identityId: "identityId",
    phoneNumber: "phoneNumber",
    primary: false,
    verified: true,
  });

  let ctx: any;
  let options: any;

  beforeEach(() => {
    ctx = {
      logger,
      repository: {
        phoneNumberRepository: {
          find: jest.fn().mockResolvedValue(phone1),
          remove: jest.fn(),
        },
      },
    };
    options = {
      identityId: "identityId",
      phoneNumber: "phoneNumber2",
    };
  });

  test("should remove phone number", async () => {
    await expect(removePhoneNumber(ctx, options)).resolves.toBeUndefined();

    expect(ctx.repository.phoneNumberRepository.remove).toHaveBeenCalledWith(phone1);
  });

  test("should throw when phone is primary", async () => {
    phone1.primary = true;

    await expect(removePhoneNumber(ctx, options)).rejects.toThrow(ClientError);
  });
});
